backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
pymysql
python-dotenv
httpx
sentence-transformers
faiss-cpu
numpy
cryptography